[YP-14] move types to props files
