+++ packages/client/src/Components/SignInForm/SignInForm.props.ts
@@ -4,3 +4,10 @@ export type SignInFormProps = {
   isLoading: boolean;
   whenSubmitForm: (userInfo: UserLogin) => void;
 };
+
+export enum TextFieldsKeys {
+  login = 'login',
+  password = 'password',
+}
+
+export type SignInFormData = Record<TextFieldsKeys, string>;
