--- packages/client/src/Components/SignUpForm/SignUp.props.ts
@@ -1,6 +1,0 @@
-import { UserCreate } from '@Types/UserCreate';
-
-export type SignUpFormProps = {
-  isLoading: boolean;
-  whenSubmitForm: (userInfo: UserCreate) => void;
-};
